requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dialogflow==0.8.0
Flask==1.1.2
google-api-core==1.16.0
google-auth==1.13.1
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
twilio==6.38.0
urllib3==1.25.8
Werkzeug==1.0.1
setuptools==46.1.3
beautifulsoup4==4.9.0
requests==2.23.0
lxml==4.5.0